+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from "./anchors";
+export * from "./anchor";
 export * from "./formatting";
 export * from "./helpers";
 export * from "./rich_list";
